Abe_Reports
Here are the updated enum classes, as well as the MVC classes for Reports. Currently the search option searches for the Reportid, but I am working on adding the other two search features to the ReadReports page. Also, the code isn't fully cleaned or commented but I wanted you to have the search by id for the heatmap and comment features. The rest  will be done by Sunday.